simpler and WCAG wording
"... without loss of formatting function or clarity."  ... huh?
-> "... without loss of content or functionality."
Matches WCAG wording exactly and seems more correct and simpler.

"...to resize or recolor text...."
-> "...to change the way text is displayed..."

Text customization is about much more than resizing or recoloring text (see http://w3c.github.io/low-vision-a11y-tf/requirements.html#user-needs ). I know the video only covers a few points for focus -- but it's good to say "for example" or "etc" to help indicate there are more. And then in the concluding statements like this, make it broader.

fyi, this is as much about "coding" as it is about design, but I think OK not to clarify that.