important wording tweak
I really had trouble with this sentence and feel strongly that it needs editing -- although I am not set on this proposed edit.

I had trouble with the phrase "audio played with visual content". I think the proposed edit is enough to clarify it -- and am also happy if others have a better idea!